詳細ページへのリンクをindexページからlabへ移動
